@@ -1,42 +1,41 @@
 module github.com/willie68/GoBlobStore
 
-go 1.18
+go 1.21
 
 require (
+	dario.cat/mergo v1.0.0
 	github.com/akgarhwal/bloomfilter v0.1.1
 	github.com/aphistic/golf v0.0.0-20180712155816-02c07f170c5a
 	github.com/blugelabs/bluge v0.2.2
 	github.com/blugelabs/query_string v0.3.0
 	github.com/drone/envsubst v1.0.3
-	github.com/go-chi/chi/v5 v5.0.8
+	github.com/go-chi/chi/v5 v5.0.10
 	github.com/go-chi/cors v1.2.1
 	github.com/go-chi/httptracer v0.3.0
-	github.com/go-chi/render v1.0.2
-	github.com/go-playground/validator/v10 v10.11.1
-	github.com/google/uuid v1.3.0
-	github.com/imdario/mergo v0.3.13
-	github.com/minio/minio-go/v7 v7.0.45
-	github.com/nsf/jsondiff v0.0.0-20210926074059-1e845ec5d249
+	github.com/go-chi/render v1.0.3
+	github.com/go-playground/validator/v10 v10.15.4
+	github.com/google/uuid v1.3.1
+	github.com/minio/minio-go/v7 v7.0.63
+	github.com/nsf/jsondiff v0.0.0-20230430225905-43f6cf3098c1
 	github.com/opentracing/opentracing-go v1.2.0
 	github.com/pkg/errors v0.9.1
-	github.com/prometheus/client_golang v1.14.0
-	github.com/shirou/gopsutil/v3 v3.22.12
+	github.com/prometheus/client_golang v1.16.0
+	github.com/samber/do v1.6.0
+	github.com/shirou/gopsutil/v3 v3.23.8
 	github.com/spf13/pflag v1.0.5
-	github.com/stretchr/testify v1.8.1
+	github.com/stretchr/testify v1.8.4
 	github.com/uber/jaeger-client-go v2.30.0+incompatible
 	github.com/vfaronov/httpheader v0.1.0
-	go.mongodb.org/mongo-driver v1.11.1
-	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+	github.com/willie68/micro-vault v0.0.0-20230914133328-9e686a0034c7
+	go.mongodb.org/mongo-driver v1.12.1
+	gopkg.in/natefinch/lumberjack.v2 v2.2.1
 	gopkg.in/yaml.v2 v2.4.0
 	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
-	github.com/BurntSushi/toml v1.2.0 // indirect
-	github.com/HdrHistogram/hdrhistogram-go v1.1.2 // indirect
 	github.com/RoaringBitmap/roaring v1.2.1 // indirect
 	github.com/ajg/form v1.5.1 // indirect
-	github.com/aphistic/sweet v0.3.0 // indirect
 	github.com/axiomhq/hyperloglog v0.0.0-20220105174342-98591331716a // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/bits-and-blooms/bitset v1.3.0 // indirect
