@@ -2,36 +2,37 @@ package config
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 
+	"dario.cat/mergo"
 	"github.com/drone/envsubst"
-	"github.com/imdario/mergo"
+	"github.com/pkg/errors"
+	"github.com/samber/do"
 	"github.com/willie68/GoBlobStore/internal/api"
+	"github.com/willie68/GoBlobStore/internal/logging"
+	"github.com/willie68/GoBlobStore/internal/services/health"
 	"gopkg.in/yaml.v3"
 )
 
-// Servicename Name of the service
+// Servicename the name of this service
 const Servicename = "goblob-service"
+
+// DoServiceConfig the name of the injected config
+const DoServiceConfig = "service_config"
 
 // Config our service configuration
 type Config struct {
-	//port of the http server
-	Port int `yaml:"port"`
-	//port of the https server
-	Sslport int `yaml:"sslport"`
-	//this is the url how to connect to this service from outside
-	ServiceURL string `yaml:"serviceURL"`
-
+	// all secrets will be stored in this file, same structure as the main config file
 	SecretFile string `yaml:"secretfile"`
 
 	Apikey bool `yaml:"apikey"`
+	// all configuration of internal services can be stored here
+	Service Service `yaml:"service"`
 
-	Logging LoggingConfig `yaml:"logging"`
-
-	HealthCheck HealthCheck `yaml:"healthcheck"`
+	// configure logging to gelf logging system
+	Logging logging.LoggingConfig `yaml:"logging"`
 
 	Auth Authentication `yaml:"auth"`
 
@@ -42,6 +43,39 @@ type Config struct {
 	OpenTracing OpenTracing `yaml:"opentracing"`
 
 	Metrics Metrics `yaml:"metrics"`
+
+	Profiling Profiling `yaml:"profiling"`
+}
+
+// Service the configuration of services inside this ms
+type Service struct {
+	HTTP HTTP `yaml:"http"`
+	// special config for health checks
+	HealthSystem health.Config `yaml:"healthcheck"`
+	// CA service will be used, microvault
+	CA CAService `yaml:"ca"`
+}
+
+// HTTP configuration of the http service
+type HTTP struct {
+	// port of the http server
+	Port int `yaml:"port"`
+	// port of the https server
+	Sslport int `yaml:"sslport"`
+	// this is the url how to connect to this service from outside
+	ServiceURL string `yaml:"serviceURL"`
+	// other dns names (used for certificate)
+	DNSNames []string `yaml:"dnss"`
+	// other ips (used for certificate)
+	IPAddresses []string `yaml:"ips"`
+}
+
+// CAService the micro-vault ca service config
+type CAService struct {
+	UseCA     bool   `yaml:"useca"`
+	URL       string `yaml:"url"`
+	AccessKey string `yaml:"accesskey"`
+	Secret    string `yaml:"secret"`
 }
 
 // Authentication configuration
@@ -75,20 +109,6 @@ type Storage struct {
 	Properties   map[string]any `yaml:"properties"`
 }
 
-// HealthCheck configuration for the health check system
-type HealthCheck struct {
-	Period int `yaml:"period"`
-}
-
-// LoggingConfig configuration for the gelf logging
-type LoggingConfig struct {
-	Level    string `yaml:"level"`
-	Filename string `yaml:"filename"`
-
-	Gelfurl  string `yaml:"gelf-url"`
-	Gelfport int    `yaml:"gelf-port"`
-}
-
 // OpenTracing configuration
 type OpenTracing struct {
 	Host     string `yaml:"host"`
@@ -100,19 +120,29 @@ type Metrics struct {
 	Enable bool `yaml:"enable"`
 }
 
+// Metrics configuration
+type Profiling struct {
+	Enable bool `yaml:"enable"`
+}
+
 var defaultHeaderMapping = map[string]string{api.TenantHeaderKey: "X-tenant", api.RetentionHeaderKey: "X-retention", api.APIKeyHeaderKey: "X-apikey", api.FilenameKey: "X-filename", api.BlobIDHeaderKey: "X-blobid", api.HeaderPrefixKey: "X-"}
 
 // DefaultConfig default configuration
 var DefaultConfig = Config{
-	Port:       8000,
-	Sslport:    8443,
-	ServiceURL: "https://127.0.0.1:8443",
-	SecretFile: "",
-	Apikey:     true,
-	HealthCheck: HealthCheck{
-		Period: 30,
+	Service: Service{
+		HTTP: HTTP{
+			Port:       8000,
+			Sslport:    8443,
+			ServiceURL: "https://127.0.0.1:8443",
+		},
+		HealthSystem: health.Config{
+			Period:     30,
+			StartDelay: 3,
+		},
 	},
-	Logging: LoggingConfig{
+	Apikey:     true,
+	SecretFile: "",
+	Logging: logging.LoggingConfig{
 		Level:    "INFO",
 		Filename: "${configdir}/logging.log",
 	},
@@ -144,6 +174,20 @@ func GetDefaultConfigFolder() (string, error) {
 	return configFolder, nil
 }
 
+// GetDefaultConfigfile getting the default config file
+func GetDefaultConfigfile() (string, error) {
+	configFolder, err := GetDefaultConfigFolder()
+	if err != nil {
+		return "", errors.Wrap(err, "can't load config file")
+	}
+	configFolder = filepath.Join(configFolder, "service")
+	err = os.MkdirAll(configFolder, os.ModePerm)
+	if err != nil {
+		return "", errors.Wrap(err, "can't load config file")
+	}
+	return filepath.Join(configFolder, "service.yaml"), nil
+}
+
 // ReplaceConfigdir replace the configdir macro
 func ReplaceConfigdir(s string) (string, error) {
 	if strings.Contains(s, "${configdir}") {
@@ -165,6 +209,11 @@ func init() {
 	config = DefaultConfig
 }
 
+// Provide provide the config to the dependency injection
+func (c *Config) Provide() {
+	do.ProvideNamedValue[Config](nil, DoServiceConfig, *c)
+}
+
 // Get returns loaded config
 func Get() Config {
 	return config
@@ -181,7 +230,7 @@ func Load() error {
 	if err != nil {
 		return err
 	}
-	data, err := ioutil.ReadFile(File)
+	data, err := os.ReadFile(File)
 	if err != nil {
 		return fmt.Errorf("can't load config file: %s", err.Error())
 	}
@@ -211,7 +260,7 @@ func Load() error {
 func readSecret() error {
 	secretFile := config.SecretFile
 	if secretFile != "" {
-		data, err := ioutil.ReadFile(secretFile)
+		data, err := os.ReadFile(secretFile)
 		if err != nil {
 			return fmt.Errorf("can't load secret file: %s", err.Error())
 		}
