@@ -6,12 +6,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
 
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/utils"
 	"github.com/willie68/GoBlobStore/pkg/model"
 )
@@ -92,21 +90,21 @@ func (s *BlobStorage) updateBlobDescriptionV2(_ string, b *model.BlobDescription
 func (s *BlobStorage) getBlobV2(id string, w io.Writer) error {
 	binFile, err := s.buildFilenameV2(id, BinaryExt)
 	if err != nil {
-		log.Logger.Errorf("error building filename: %v", err)
+		logger.Errorf("error building filename: %v", err)
 		return err
 	}
 	if _, err := os.Stat(binFile); os.IsNotExist(err) {
-		log.Logger.Errorf("error not exists: %v", err)
+		logger.Errorf("error not exists: %v", err)
 		return os.ErrNotExist
 	}
 	f, err := os.Open(binFile)
 	if err != nil {
-		log.Logger.Errorf("error opening file: %v", err)
+		logger.Errorf("error opening file: %v", err)
 		return err
 	}
 	defer f.Close()
 	if _, err = io.Copy(w, f); err != nil {
-		log.Logger.Errorf("error on copy: %v", err)
+		logger.Errorf("error on copy: %v", err)
 		return err
 	}
 	return nil
@@ -135,7 +133,6 @@ func (s *BlobStorage) storeBlobV2(b *model.BlobDescription, f io.Reader) (string
 	s.cm.Lock()
 	defer s.cm.Unlock()
 	s.bdCch[b.BlobID] = *b
-	//	go s.buildHash(b.BlobID)
 	return b.BlobID, nil
 }
 
@@ -159,7 +156,6 @@ func (s *BlobStorage) writeBinFileV2(id string, r io.Reader) (int64, string, err
 
 	size, err := io.Copy(w, r)
 
-	//size, err := f.ReadFrom(r)
 	if err != nil {
 		_ = f.Close()
 		_ = os.Remove(binFile)
@@ -200,7 +196,7 @@ func (s *BlobStorage) writeJSONFileV2(b *model.BlobDescription) error {
 		return err
 	}
 
-	err = ioutil.WriteFile(jsonFile, jsn, os.ModePerm)
+	err = os.WriteFile(jsonFile, jsn, os.ModePerm)
 	if err != nil {
 		_ = os.Remove(jsonFile)
 		return err
@@ -220,7 +216,7 @@ func (s *BlobStorage) writeRetentionFile(b *model.BlobDescription) error {
 		return err
 	}
 
-	err = ioutil.WriteFile(jsonFile, jsn, os.ModePerm)
+	err = os.WriteFile(jsonFile, jsn, os.ModePerm)
 	if err != nil {
 		_ = os.Remove(jsonFile)
 		return err
@@ -236,13 +232,13 @@ func (s *BlobStorage) getRetention(id string) (*model.RetentionEntry, error) {
 	}
 	dat, err := os.ReadFile(jsonFile)
 	if err != nil {
-		log.Logger.Errorf("GetRetention: error getting file data for: %s\r\n%v", jsonFile, err)
+		logger.Errorf("GetRetention: error getting file data for: %s\r\n%v", jsonFile, err)
 		return nil, err
 	}
 	r := model.RetentionEntry{}
 	err = json.Unmarshal(dat, &r)
 	if err != nil {
-		log.Logger.Errorf("GetRetention: deserialization error: %s\r\n%v", jsonFile, err)
+		logger.Errorf("GetRetention: deserialization error: %s\r\n%v", jsonFile, err)
 		return nil, err
 	}
 	return &r, nil
