@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"github.com/willie68/GoBlobStore/internal/api"
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	services "github.com/willie68/GoBlobStore/internal/services"
 
 	"github.com/go-chi/chi/v5"
@@ -14,9 +13,7 @@ import (
 	"github.com/willie68/GoBlobStore/internal/utils/httputils"
 )
 
-/*
-AdminRoutes getting all routes for the admin endpoint
-*/
+// AdminRoutes getting all routes for the admin endpoint
 func AdminRoutes() (string, *chi.Mux) {
 	router := chi.NewRouter()
 	router.With(api.RoleCheck([]api.Role{api.RoleTenantAdmin})).Get("/check", GetCheck)
@@ -76,7 +73,7 @@ func PostCheck(response http.ResponseWriter, request *http.Request) {
 		httputils.Err(response, request, serror.BadRequest(nil, "missing-tenant", msg))
 		return
 	}
-	log.Logger.Infof("do check for tenant %s", tenant)
+	logger.Infof("do check for tenant %s", tenant)
 	cMan, err := services.GetMigrationManagement()
 	if err != nil {
 		httputils.Err(response, request, serror.InternalServerError(err))
@@ -118,7 +115,7 @@ func PostRestore(response http.ResponseWriter, request *http.Request) {
 		httputils.Err(response, request, serror.BadRequest(nil, "missing-tenant", msg))
 		return
 	}
-	log.Logger.Infof("do restore for tenant %s", tenant)
+	logger.Infof("do restore for tenant %s", tenant)
 	rMan, err := services.GetMigrationManagement()
 	if err != nil {
 		httputils.Err(response, request, serror.InternalServerError(err))
