@@ -15,7 +15,6 @@ import (
 	"github.com/vfaronov/httpheader"
 	"github.com/willie68/GoBlobStore/internal/api"
 	"github.com/willie68/GoBlobStore/internal/config"
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/serror"
 	services "github.com/willie68/GoBlobStore/internal/services"
 	"github.com/willie68/GoBlobStore/internal/services/interfaces"
@@ -72,11 +71,9 @@ func getBlobLocation(blobid string) string {
 	return fmt.Sprintf(BaseURL+blobsSubpath+"/%s", blobid)
 }
 
-/*
-GetBlob getting one blob file for a tenant from the storage
-path parameter
-id: the id of the blob file
-*/
+// GetBlob getting one blob file for a tenant from the storage
+// path parameter
+// id: the id of the blob file
 func GetBlob(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -149,11 +146,9 @@ func GetBlob(response http.ResponseWriter, request *http.Request) {
 	}
 }
 
-/*
-GetBlobInfo getting the info of a blob file from the storage
-path param:
-id: the id of the blob file
-*/
+// GetBlobInfo getting the info of a blob file from the storage
+// path param:
+// id: the id of the blob file
 func GetBlobInfo(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -193,11 +188,9 @@ func GetBlobInfo(response http.ResponseWriter, request *http.Request) {
 	render.JSON(response, request, b)
 }
 
-/*
-PutBlobInfo getting the info of a blob file from the storage
-path param:
-id: the id of the blob file
-*/
+// PutBlobInfo getting the info of a blob file from the storage
+// path param:
+// id: the id of the blob file
 func PutBlobInfo(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -257,11 +250,9 @@ func PutBlobInfo(response http.ResponseWriter, request *http.Request) {
 	render.JSON(response, request, b)
 }
 
-/*
-GetBlobResetRetention restting the retention time of a blob to the new value
-path param:
-id: the id of the lob file
-*/
+// GetBlobResetRetention restting the retention time of a blob to the new value
+// path param:
+// id: the id of the lob file
 func GetBlobResetRetention(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -300,12 +291,10 @@ func GetBlobResetRetention(response http.ResponseWriter, request *http.Request) 
 	render.JSON(response, request, found)
 }
 
-/*
-GetBlobs query all blobs from the storage for a tenant
-query params
-offset: the offset to start from
-limit: max count of blobs
-*/
+// GetBlobs query all blobs from the storage for a tenant
+// query params
+// offset: the offset to start from
+// limit: max count of blobs
 func GetBlobs(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -497,21 +486,26 @@ func getFilename(header http.Header) (string, error) {
 
 	filenameHeader, ok := config.Get().HeaderMapping[api.FilenameKey]
 	if ok {
-		filename = header.Get(filenameHeader)
-		header, _, err := httpheader.DecodeExtValue(filename)
-		if err != nil {
-			return "", err
+		fn := header.Get(filenameHeader)
+		if fn != "" {
+			// check if there is an RFC 8187 coded value
+			if strings.ContainsAny(fn, "'") {
+				decodedValue, _, err := httpheader.DecodeExtValue(fn)
+				if err != nil {
+					return "", err
+				}
+				filename = decodedValue
+			} else {
+				filename = fn
+			}
 		}
-		filename = header
 	}
 	return filename, nil
 }
 
-/*
-DeleteBlob delete a dedicated blob from the storage for the tenant
-path param
-id: the id of the blob to remove
-*/
+// DeleteBlob delete a dedicated blob from the storage for the tenant
+// path param
+// id: the id of the blob to remove
 func DeleteBlob(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -555,13 +549,11 @@ func DeleteBlob(response http.ResponseWriter, request *http.Request) {
 	render.JSON(response, request, idStr)
 }
 
-/*
-SearchBlobs search for blobs meeting the criteria
-query params
-offset: the offset to start from
-limit: max count of blobs
-q: query to use
-*/
+// SearchBlobs search for blobs meeting the criteria
+// query params
+// offset: the offset to start from
+// limit: max count of blobs
+// q: query to use
 func SearchBlobs(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -599,7 +591,7 @@ func SearchBlobs(response http.ResponseWriter, request *http.Request) {
 	}
 	query := string(b)
 	if query != "" {
-		log.Logger.Debugf("search for blobs with: %s", query)
+		logger.Debugf("search for blobs with: %s", query)
 	}
 	blobs := make([]string, 0)
 	index := 0
@@ -620,11 +612,9 @@ func SearchBlobs(response http.ResponseWriter, request *http.Request) {
 	render.JSON(response, request, blobs)
 }
 
-/*
-GetBlobCheck getting the latest check info of a blob file from the storage
-path param:
-id: the id of the blob file
-*/
+// GetBlobCheck getting the latest check info of a blob file from the storage
+// path param:
+// id: the id of the blob file
 func GetBlobCheck(response http.ResponseWriter, request *http.Request) {
 	tenant, err := httputils.TenantID(request)
 	if err != nil {
@@ -689,7 +679,7 @@ func PostBlobCheck(response http.ResponseWriter, request *http.Request) {
 	}
 	idStr := chi.URLParam(request, "id")
 
-	log.Logger.Infof("do check for tenant %s on blob %s", tenant, idStr)
+	logger.Infof("do check for tenant %s on blob %s", tenant, idStr)
 	storage, err := getTenantStore(tenant)
 	if err != nil {
 		httputils.Err(response, request, serror.InternalServerError(err))
