@@ -10,7 +10,6 @@ import (
 	"github.com/go-chi/render"
 	"github.com/willie68/GoBlobStore/internal/api"
 	"github.com/willie68/GoBlobStore/internal/config"
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/serror"
 	services "github.com/willie68/GoBlobStore/internal/services"
 	"github.com/willie68/GoBlobStore/internal/services/factory"
@@ -19,9 +18,7 @@ import (
 	"github.com/willie68/GoBlobStore/pkg/model"
 )
 
-/*
-ConfigRoutes getting all routes for the config endpoint
-*/
+// ConfigRoutes getting all routes for the config endpoint
 func ConfigRoutes() (string, *chi.Mux) {
 	router := chi.NewRouter()
 	router.With(api.RoleCheck([]api.Role{api.RoleAdmin})).Post("/", PostCreateTenant)
@@ -31,9 +28,7 @@ func ConfigRoutes() (string, *chi.Mux) {
 	return BaseURL + configSubpath, router
 }
 
-/*
-StoresRoutes getting all routes for the stores endpoint, this is part of the new api. But manly here only a new name.
-*/
+// StoresRoutes getting all routes for the stores endpoint, this is part of the new api. But manly here only a new name.
 func StoresRoutes() (string, *chi.Mux) {
 	router := chi.NewRouter()
 	router.With(api.RoleCheck([]api.Role{api.RoleAdmin})).Post("/", PostCreateTenant)
@@ -115,7 +110,7 @@ func PostCreateTenant(response http.ResponseWriter, request *http.Request) {
 		httputils.Err(response, request, serror.BadRequest(nil, "missing-tenant", msg))
 		return
 	}
-	log.Logger.Infof("create store for tenant %s", tenant)
+	logger.Infof("create store for tenant %s", tenant)
 	tntsrv, err := services.GetTenantSrv()
 	if err != nil {
 		httputils.Err(response, request, serror.InternalServerError(err))
@@ -134,14 +129,14 @@ func PostCreateTenant(response http.ResponseWriter, request *http.Request) {
 		return
 	}
 
-	rsp := model.CreateResponse{
-		TenantID: tenant,
-	}
-
 	if !config.Get().Engine.AllowTntBackup && cfg.Storageclass != "" {
 		err := errors.New("tenant base backups are not allowed")
 		httputils.Err(response, request, serror.BadRequest(err))
 		return
+	}
+
+	rsp := model.CreateResponse{
+		TenantID: tenant,
 	}
 
 	if config.Get().Engine.AllowTntBackup && cfg.Storageclass != "" {
@@ -163,7 +158,11 @@ func PostCreateTenant(response http.ResponseWriter, request *http.Request) {
 			httputils.Err(response, request, serror.InternalServerError(err))
 			return
 		}
-		stf.RemoveStorage(tenant)
+		err = stf.RemoveStorage(tenant)
+		if err != nil {
+			httputils.Err(response, request, serror.InternalServerError(err))
+			return
+		}
 		rsp.Backup = cfg.Storageclass
 	}
 
