@@ -6,22 +6,31 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+const (
+	versionStr    = "1.2.3-special"
+	msgErrParsing = "Error in parsing"
+)
+
 func TestParsing(t *testing.T) {
-	versionStr := "1.2.3-special"
+	ast := assert.New(t)
 
 	version, err := ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
-	assert.Equal(t, 1, version.Major, "Major version not equal")
-	assert.Equal(t, 2, version.Minor, "Minor version not equal")
-	assert.Equal(t, 3, version.Patch, "Patch version not equal")
-	assert.Equal(t, "special", version.Special, "Special version not equal")
+	ast.Nil(err)
+	ast.Equal(1, version.Major, "Major version not equal")
+	ast.Equal(2, version.Minor, "Minor version not equal")
+	ast.Equal(3, version.Patch, "Patch version not equal")
+	ast.Equal("special", version.Special, "Special version not equal")
 }
 
 func TestCompareEqual(t *testing.T) {
-	versionStr := "1.2.3-special"
+	ast := assert.New(t)
 
 	version, err := ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
+	ast.Nil(err, msgErrParsing)
+
+	ast.False(version.IsEqual(Version{
+		Major: 2,
+	}))
 
 	version2 := Version{
 		Major:   1,
@@ -30,14 +39,20 @@ func TestCompareEqual(t *testing.T) {
 		Special: "special",
 	}
 
-	assert.False(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.True(t, version.IsEqual(version2))
+	ast.False(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.True(version.IsEqual(version2))
+	ast.False(version.IsEqual(Version{
+		Major:   1,
+		Minor:   2,
+		Patch:   3,
+		Special: "Willie",
+	}))
 
-	versionStr = "1.2.3"
+	otherVersionStr := "1.2.3"
 
-	version, err = ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
+	version, err = ParseVersion(otherVersionStr)
+	ast.Nil(err, msgErrParsing)
 
 	version2 = Version{
 		Major: 1,
@@ -45,43 +60,56 @@ func TestCompareEqual(t *testing.T) {
 		Patch: 3,
 	}
 
-	assert.False(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.True(t, version.IsEqual(version2))
+	ast.False(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.True(version.IsEqual(version2))
 
-	versionStr = "1.2"
+	otherVersionStr = "1.2"
 
-	version, err = ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
+	version, err = ParseVersion(otherVersionStr)
+	ast.Nil(err, msgErrParsing)
 
 	version2 = Version{
 		Major: 1,
 		Minor: 2,
 	}
 
-	assert.False(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.True(t, version.IsEqual(version2))
+	ast.False(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.True(version.IsEqual(version2))
 
-	versionStr = "1"
+	otherVersionStr = "1"
 
-	version, err = ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
+	version, err = ParseVersion(otherVersionStr)
+	ast.Nil(err, msgErrParsing)
 
 	version2 = Version{
 		Major: 1,
 	}
 
-	assert.False(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.True(t, version.IsEqual(version2))
+	ast.False(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.True(version.IsEqual(version2))
 }
 
 func TestCompareGreater(t *testing.T) {
-	versionStr := "1.2.3-special"
+	ast := assert.New(t)
 
 	version, err := ParseVersion(versionStr)
-	assert.Nil(t, err, "Error in parsing")
+	ast.Nil(err, msgErrParsing)
+
+	ast.True(version.IsGreaterThan(Version{
+		Major: 0,
+	}))
+	ast.False(version.IsGreaterThan(Version{
+		Major: 2,
+	}))
+	ast.False(version.IsSmallerThan(Version{
+		Major: 0,
+	}))
+	ast.True(version.IsSmallerThan(Version{
+		Major: 2,
+	}))
 
 	version2 := Version{
 		Major: 1,
@@ -89,30 +117,48 @@ func TestCompareGreater(t *testing.T) {
 		Patch: 2,
 	}
 
-	assert.True(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.False(t, version2.IsGreaterThan(version))
-	assert.True(t, version2.IsSmallerThan(version))
-	assert.False(t, version.IsEqual(version2))
+	ast.True(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.False(version2.IsGreaterThan(version))
+	ast.True(version2.IsSmallerThan(version))
+	ast.False(version.IsEqual(version2))
 
 	version2 = Version{
 		Major: 1,
 		Minor: 2,
 	}
 
-	assert.True(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.False(t, version2.IsGreaterThan(version))
-	assert.True(t, version2.IsSmallerThan(version))
-	assert.False(t, version.IsEqual(version2))
+	ast.True(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.False(version2.IsGreaterThan(version))
+	ast.True(version2.IsSmallerThan(version))
+	ast.False(version.IsEqual(version2))
 
 	version2 = Version{
 		Major: 1,
 	}
 
-	assert.True(t, version.IsGreaterThan(version2))
-	assert.False(t, version.IsSmallerThan(version2))
-	assert.False(t, version2.IsGreaterThan(version))
-	assert.True(t, version2.IsSmallerThan(version))
-	assert.False(t, version.IsEqual(version2))
+	ast.True(version.IsGreaterThan(version2))
+	ast.False(version.IsSmallerThan(version2))
+	ast.False(version2.IsGreaterThan(version))
+	ast.True(version2.IsSmallerThan(version))
+	ast.False(version.IsEqual(version2))
+}
+
+func TestVersionString(t *testing.T) {
+	ast := assert.New(t)
+
+	version := Version{
+		Major: 1,
+	}
+	ast.Equal("1.0.0", version.String())
+
+	version.Minor = 9
+	ast.Equal("1.9.0", version.String())
+
+	version.Patch = 35
+	ast.Equal("1.9.35", version.String())
+
+	version.Special = "Willie"
+	ast.Equal("1.9.35-Willie", version.String())
 }
