@@ -1,10 +1,8 @@
 // Package business the package contains the structs for the business rules of the storage system
 package business
 
-/*
-This type doing all the business logic of storing blobs of the service.
-Managing backup and cache requests, managing the Retentions
-*/
+// This type doing all the business logic of storing blobs of the service.
+// Managing backup and cache requests, managing the Retentions
 import (
 	"errors"
 	"fmt"
@@ -12,7 +10,6 @@ import (
 	"runtime"
 	"time"
 
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/services/interfaces"
 	"github.com/willie68/GoBlobStore/pkg/model"
 )
@@ -28,6 +25,7 @@ type MainStorage struct {
 	CchSrv      interfaces.BlobStorage
 	IdxSrv      interfaces.Index
 	TntBckSrv   interfaces.BlobStorage
+	TntMgr      interfaces.TenantManager
 	Bcksyncmode bool
 	Tenant      string
 	hasIdx      bool
@@ -94,6 +92,7 @@ func (m *MainStorage) StoreBlob(b *model.BlobDescription, f io.Reader) (string, 
 		}
 	}
 	go m.cacheFile(b)
+	go m.addStorageSize(id)
 	gor := (runtime.NumGoroutine() / 1000)
 	time.Sleep(time.Duration(gor) * time.Millisecond)
 	return id, err
@@ -131,13 +130,13 @@ func (m *MainStorage) cacheFileByID(id string) {
 	if m.CchSrv != nil {
 		ok, err := m.CchSrv.HasBlob(id)
 		if err != nil {
-			log.Logger.Errorf("main: cacheFileByID: check blob: %s, %v", id, err)
+			logger.Errorf("main: cacheFileByID: check blob: %s, %v", id, err)
 			return
 		}
 		if !ok {
 			b, err := m.GetBlobDescription(id)
 			if err != nil {
-				log.Logger.Errorf("main: cacheFileByID: getDescription: %s, %v", id, err)
+				logger.Errorf("main: cacheFileByID: getDescription: %s, %v", id, err)
 				return
 			}
 			m.cacheFile(b)
@@ -149,7 +148,7 @@ func (m *MainStorage) cacheFile(b *model.BlobDescription) {
 	if m.CchSrv != nil {
 		ok, err := m.CchSrv.HasBlob(b.BlobID)
 		if err != nil {
-			log.Logger.Errorf("main: cacheFile: check blob: %s, %v", b.BlobID, err)
+			logger.Errorf("main: cacheFile: check blob: %s, %v", b.BlobID, err)
 			return
 		}
 		if !ok {
@@ -157,13 +156,13 @@ func (m *MainStorage) cacheFile(b *model.BlobDescription) {
 			go func() {
 				defer wr.Close()
 				if err := m.StgSrv.RetrieveBlob(b.BlobID, wr); err != nil {
-					log.Logger.Errorf("main: cacheFile: retrieve, error getting blob: %s, %v", b.BlobID, err)
+					logger.Errorf("main: cacheFile: retrieve, error getting blob: %s, %v", b.BlobID, err)
 				}
 				// close the writer, so the reader knows there's no more data
 			}()
 			defer rd.Close()
 			if _, err := m.CchSrv.StoreBlob(b, rd); err != nil {
-				log.Logger.Errorf("main: cacheFile: store, error getting blob: %s, %v", b.BlobID, err)
+				logger.Errorf("main: cacheFile: store, error getting blob: %s, %v", b.BlobID, err)
 			}
 		}
 	}
@@ -185,7 +184,7 @@ func (m *MainStorage) bckFile(srv interfaces.BlobStorage, b *model.BlobDescripti
 	if srv != nil {
 		ok, err := srv.HasBlob(b.BlobID)
 		if err != nil {
-			log.Logger.Errorf("main: backupFile: check blob: %s, %v", b.BlobID, err)
+			logger.Errorf("main: backupFile: check blob: %s, %v", b.BlobID, err)
 			return
 		}
 		if !ok {
@@ -194,12 +193,12 @@ func (m *MainStorage) bckFile(srv interfaces.BlobStorage, b *model.BlobDescripti
 				// close the writer, so the reader knows there's no more data
 				defer wr.Close()
 				if err := m.StgSrv.RetrieveBlob(id, wr); err != nil {
-					log.Logger.Errorf("main: backupFile: retrieve, error getting blob: %s, %v", id, err)
+					logger.Errorf("main: backupFile: retrieve, error getting blob: %s, %v", id, err)
 				}
 			}()
 			defer rd.Close()
 			if _, err := srv.StoreBlob(b, rd); err != nil {
-				log.Logger.Errorf("main: backupFile: store, error getting blob: %s, %v", id, err)
+				logger.Errorf("main: backupFile: store, error getting blob: %s, %v", id, err)
 			}
 		}
 	}
@@ -210,7 +209,7 @@ func (m *MainStorage) restoreFile(b *model.BlobDescription) {
 		id := b.BlobID
 		ok, err := m.BckSrv.HasBlob(id)
 		if err != nil {
-			log.Logger.Errorf("main: restoreFile: check blob: %s, %v", id, err)
+			logger.Errorf("main: restoreFile: check blob: %s, %v", id, err)
 			return
 		}
 		if ok {
@@ -219,14 +218,34 @@ func (m *MainStorage) restoreFile(b *model.BlobDescription) {
 				// close the writer, so the reader knows there's no more data
 				defer wr.Close()
 				if err := m.BckSrv.RetrieveBlob(id, wr); err != nil {
-					log.Logger.Errorf("main: restoreFile: retrieve, error getting blob: %s, %v", id, err)
+					logger.Errorf("main: restoreFile: retrieve, error getting blob: %s, %v", id, err)
 				}
 			}()
 			defer rd.Close()
 			if _, err := m.StgSrv.StoreBlob(b, rd); err != nil {
-				log.Logger.Errorf("main: restoreFile: store, error getting blob: %s, %v", id, err)
+				logger.Errorf("main: restoreFile: store, error getting blob: %s, %v", id, err)
 			}
+			go m.cacheFileByID(id)
 		}
+	}
+}
+
+// addStorageSize adjust the storage size for the tenant
+func (m *MainStorage) addStorageSize(id string) {
+	bd, err := m.GetBlobDescription(id)
+	if err != nil {
+		logger.Errorf("adjust: can't get blob description: %v", err)
+		return
+	}
+	if m.TntMgr != nil {
+		m.TntMgr.AddSize(m.Tenant, bd.ContentLength)
+	}
+}
+
+// subStorageSize subtract the storage size for the tenant
+func (m *MainStorage) subStorageSize(bd *model.BlobDescription) {
+	if m.TntMgr != nil {
+		m.TntMgr.SubSize(m.Tenant, bd.ContentLength)
 	}
 }
 
@@ -299,8 +318,6 @@ func (m *MainStorage) RetrieveBlob(id string, w io.Writer) error {
 		}
 		return err
 	}
-
-	go m.cacheFileByID(id)
 	return nil
 }
 
@@ -321,13 +338,18 @@ func (m *MainStorage) retrieveFromCache(id string, w io.Writer) bool {
 
 // DeleteBlob removing a blob from the storage system
 func (m *MainStorage) DeleteBlob(id string) error {
-	err := m.StgSrv.DeleteBlob(id)
+	bd, err := m.StgSrv.GetBlobDescription(id)
 	if err != nil {
 		return err
 	}
+	err = m.StgSrv.DeleteBlob(id)
+	if err != nil {
+		return err
+	}
+	go m.subStorageSize(bd)
 	if m.BckSrv != nil {
 		if err = m.BckSrv.DeleteBlob(id); err != nil {
-			log.Logger.Errorf("error deleting blob on backup: %v", err)
+			logger.Errorf("error deleting blob on backup: %v", err)
 		}
 	}
 	if m.RtnMng != nil {
@@ -335,7 +357,7 @@ func (m *MainStorage) DeleteBlob(id string) error {
 	}
 	if m.CchSrv != nil {
 		if err = m.CchSrv.DeleteBlob(id); err != nil {
-			log.Logger.Errorf("error deleting blob on cache: %v", err)
+			logger.Errorf("error deleting blob on cache: %v", err)
 		}
 	}
 	return nil
@@ -383,11 +405,11 @@ func (m *MainStorage) CheckBlob(id string) (*model.CheckInfo, error) {
 func (m *MainStorage) checkBck(id string, ri *model.Check, bd *model.BlobDescription) {
 	bckDI, err := m.BckSrv.CheckBlob(id)
 	if err != nil {
-		log.Logger.Errorf("error checking blob on backup: %v", err)
+		logger.Errorf("error checking blob on backup: %v", err)
 	}
 	bckBd, err := m.BckSrv.GetBlobDescription(id)
 	if err != nil {
-		log.Logger.Errorf("error getting blob description on backup: %v", err)
+		logger.Errorf("error getting blob description on backup: %v", err)
 	}
 	// merge stgCI and bckCI
 	ri.Backup = bckDI
@@ -423,7 +445,7 @@ func (m *MainStorage) AddRetention(r *model.RetentionEntry) error {
 	err := m.StgSrv.AddRetention(r)
 	if m.BckSrv != nil {
 		if err1 := m.BckSrv.AddRetention(r); err1 != nil {
-			log.Logger.Errorf("error adding retention on backup: %v", err1)
+			logger.Errorf("error adding retention on backup: %v", err1)
 		}
 	}
 	return err
@@ -439,7 +461,7 @@ func (m *MainStorage) DeleteRetention(id string) error {
 	err := m.StgSrv.DeleteRetention(id)
 	if m.BckSrv != nil {
 		if err1 := m.BckSrv.DeleteRetention(id); err1 != nil {
-			log.Logger.Errorf("error deleting retention on backup:%s, %v", id, err1)
+			logger.Errorf("error deleting retention on backup:%s, %v", id, err1)
 		}
 	}
 	return err
@@ -450,7 +472,7 @@ func (m *MainStorage) ResetRetention(id string) error {
 	err := m.StgSrv.ResetRetention(id)
 	if m.BckSrv != nil {
 		if err1 := m.BckSrv.ResetRetention(id); err1 != nil {
-			log.Logger.Errorf("error resetting retention on backup:%s, %v", id, err1)
+			logger.Errorf("error resetting retention on backup:%s, %v", id, err1)
 		}
 	}
 	return err
@@ -466,7 +488,7 @@ func (m *MainStorage) Close() error {
 	err := m.StgSrv.Close()
 	if m.BckSrv != nil {
 		if err1 := m.BckSrv.Close(); err1 != nil {
-			log.Logger.Errorf("error closing backup storage: %v", err1)
+			logger.Errorf("error closing backup storage: %v", err1)
 		}
 	}
 	return err
