@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/akgarhwal/bloomfilter/bloomfilter"
-	log "github.com/willie68/GoBlobStore/internal/logging"
+	"github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/services/interfaces"
 	"github.com/willie68/GoBlobStore/internal/utils"
 	"github.com/willie68/GoBlobStore/pkg/model"
@@ -30,6 +30,8 @@ const (
 var (
 	errEmptyIndex     = errors.New("empty id not allowed")
 	errNotImplemented = errors.New("method not implemented in fastcache")
+
+	logger = logging.New().WithName("fastcache")
 )
 
 // FastCache a fast cache implementation using a mix of memory and fast ssd storage
@@ -140,16 +142,16 @@ func (f *FastCache) GetBlobs(callback func(id string) bool) error {
 func (f *FastCache) StoreBlob(b *model.BlobDescription, r io.Reader) (string, error) {
 	ok, err := f.HasBlob(b.BlobID)
 	if err != nil {
-		log.Logger.Errorf("cache: error checking file: %v", err)
+		logger.Errorf("cache: error checking file: %v", err)
 		return "", err
 	}
 	if ok {
-		log.Logger.Errorf("cache: file exists")
+		logger.Errorf("cache: file exists")
 		return b.BlobID, os.ErrExist
 	}
 	size, dat, err := f.writeBinFile(b.BlobID, r)
 	if err != nil {
-		log.Logger.Errorf("cache: writing file: %v", err)
+		logger.Errorf("cache: writing file: %v", err)
 		return "", err
 	}
 	atomic.AddInt64(&f.size, size)
@@ -160,13 +162,12 @@ func (f *FastCache) StoreBlob(b *model.BlobDescription, r io.Reader) (string, er
 	})
 	for {
 		id := f.entries.HandleContrains()
-		if id != "" {
-			err = f.DeleteBlob(id)
-			if err != nil {
-				log.Logger.Errorf("cache: can't delete blob %s: %v", id, err)
-			}
-		} else {
+		if id == "" {
 			break
+		}
+		err = f.DeleteBlob(id)
+		if err != nil {
+			logger.Errorf("cache: can't delete blob %s: %v", id, err)
 		}
 	}
 	f.updateBloom(b.BlobID)
@@ -177,11 +178,11 @@ func (f *FastCache) StoreBlob(b *model.BlobDescription, r io.Reader) (string, er
 func (f *FastCache) UpdateBlobDescription(id string, b *model.BlobDescription) error {
 	ok, err := f.HasBlob(b.BlobID)
 	if err != nil {
-		log.Logger.Errorf("cache: error checking file: %v", err)
+		logger.Errorf("cache: error checking file: %v", err)
 		return err
 	}
 	if !ok {
-		log.Logger.Debugf("cache: file not exists: %s", id)
+		logger.Debugf("cache: file not exists: %s", id)
 		return nil
 	}
 	if f.inBloom(id) {
@@ -319,7 +320,7 @@ func (f *FastCache) DeleteBlob(id string) error {
 			if lid != "" {
 				err := f.deleteBlobFile(lid)
 				if err != nil {
-					log.Logger.Errorf("error deleting file: %v", err)
+					logger.Errorf("error deleting file: %v", err)
 					return err
 				}
 				f.bfDirty = true
