@@ -3,16 +3,22 @@ package services
 import (
 	"errors"
 
+	"github.com/samber/do"
 	"github.com/willie68/GoBlobStore/internal/services/business"
 	"github.com/willie68/GoBlobStore/internal/services/migration"
 
 	"github.com/willie68/GoBlobStore/internal/config"
-	log "github.com/willie68/GoBlobStore/internal/logging"
 	"github.com/willie68/GoBlobStore/internal/services/factory"
 	"github.com/willie68/GoBlobStore/internal/services/interfaces"
 )
 
-// test for interface compatibility
+// all constant definitions for the different services
+const (
+	DoTntSrv = "tntsrv"
+	DoRtnMgr = "rtnmgr"
+	DoStgf   = "stgf"
+	DoMigMgr = "migmgr"
+)
 
 var tntsrv interfaces.TenantManager
 var rtnMgr interfaces.RetentionManager
@@ -20,7 +26,7 @@ var cnfg config.Engine
 var stgf interfaces.StorageFactory
 var migMan *migration.Management
 
-// Init initialise the storage factory
+// Init initialize the storage factory
 func Init(storage config.Engine) error {
 	cnfg = storage
 	if cnfg.Storage.Storageclass == "" {
@@ -32,16 +38,20 @@ func Init(storage config.Engine) error {
 	if err != nil {
 		return err
 	}
+
 	if cnfg.Backup.Storageclass != "" {
 		bktsrv, err = factory.CreateTenantManager(cnfg.Backup)
 		if err != nil {
 			return err
 		}
 	}
+
 	tntsrv = &business.MainTenant{
 		TntSrv: tntMgr,
 		BckSrv: bktsrv,
 	}
+
+	do.ProvideNamedValue[interfaces.TenantManager](nil, DoTntSrv, tntsrv)
 
 	if cnfg.RetentionManager == "" {
 		return errors.New("no retention class given")
@@ -57,10 +67,14 @@ func Init(storage config.Engine) error {
 		return err
 	}
 
+	do.ProvideNamedValue[interfaces.RetentionManager](nil, DoRtnMgr, rtnMgr)
+
 	err = stgf.Init(storage, rtnMgr)
 	if err != nil {
 		return err
 	}
+
+	do.ProvideNamedValue[interfaces.StorageFactory](nil, DoStgf, stgf)
 
 	err = rtnMgr.Init(stgf)
 	if err != nil {
@@ -80,6 +94,9 @@ func Init(storage config.Engine) error {
 	if err != nil {
 		return err
 	}
+
+	do.ProvideNamedValue[*migration.Management](nil, DoMigMgr, migMan)
+
 	return nil
 }
 
@@ -111,21 +128,21 @@ func GetMigrationManagement() (*migration.Management, error) {
 func Close() {
 	err := stgf.Close()
 	if err != nil {
-		log.Logger.Errorf("error closing storage factory:\r\n%v,", err)
+		logger.Errorf("error closing storage factory:\r\n%v,", err)
 	}
 
 	err = rtnMgr.Close()
 	if err != nil {
-		log.Logger.Errorf("error closing retention manager:\r\n%v,", err)
+		logger.Errorf("error closing retention manager:\r\n%v,", err)
 	}
 
 	err = tntsrv.Close()
 	if err != nil {
-		log.Logger.Errorf("error closing tenant service:\r\n%v,", err)
+		logger.Errorf("error closing tenant service:\r\n%v,", err)
 	}
 
 	err = migMan.Close()
 	if err != nil {
-		log.Logger.Errorf("error closing check management:\r\n%v,", err)
+		logger.Errorf("error closing check management:\r\n%v,", err)
 	}
 }
